requirements.txt
numpy==1.26.4
pandas==2.2.2
streamlit==1.33.0
openai==1.23.1
python-dotenv==1.0.1
seaborn==0.13.2
matplotlib==3.8.4
langchain==0.1.16
langchain-experimental==0.0.57
langchain-openai==0.1.3
tabulate==0.9.0